requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
statsmodels
prophet
google-generativeai
python-dotenv
openpyxl
fpdf2
markdown
